split parent and child components
