+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
-//import HomePageForParent from './components/HomePageForParent';
-import HomePageForChildren from './components/HomePageForChildren';
+//import HomePageForParent from './components/Parent/HomePageForParent';
+import HomePageForChildren from './components/Child/HomePageForChildren';
 import './basic.scss';
 
 function App() {
